+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes that may have an impact on backwards compatibility (i.e. they may break existing workflows) are marked with `[BC]`.
 
 ## Current development (master)
+- Small speed optimization (~ 10%) to `Gitify build --force` (#78)
 - Passing arguments to `Gitify build` and  `Gitify extract` now restricts building/extracting to those specific data partitions/folders (#51, #26)
 - Escape password in backup/restore commands so they work with special characters in the password